Fixed PopupMenus and integration of PopupMenus inside of PopupScrollArea

+++ core/src/com/desitum/castleWars/libraries/menu/PopupWidget.java
@@ -164,7 +164,7 @@ public abstract class PopupWidget extends Sprite {
         return enabled;
     }
 
-    public void updateScroll(int amount, Vector3 mousePos) {
+    public void updateScroll(float amount, Vector3 mousePos) {
 
     }
